split up query-payload-types
